Add BitMEX. (#299)
This implementation is not 100% ideal; BitMEX differs
from most markets in that we have a dynamic list of futures
expiries that are different from symbol to symbol.

In order to support this, the data persisted to Context
would need to include a parsed futures expiry, and we would
also need to support dynamic expiries per market. This is
possible, but may cause problems upstream.

In that interest, this commit presents the minimal changes
needed for usable BitMEX support.